+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update `App` initialization to Svelte 5 using `mount` function
 
+### Fixed
+
+- Correct textarea closing tag in `TextArea.svelte`
+- Correct option tag in `Option.svelte`
+
 ### Dependencies
 
 - Bumped `svelte` to `5.19.7`